Implement analysis mode
The XBoard commands 'analyze' and 'exit' are recognized, to set and
clear a new analysis flag (and set force mode). While the flag is set
the input loop will order a background search on the current position
while waiting for input. (Note that the 'exit' command outside XBoard
mode is synonymous for 'quit'!)
BEWARE: automatic updates interrupt the analysis.