Allow passing command to InputCommand
InputCommand was only called with argument command=NULL, but written such
that it would execute the passed command rather than read one from stdin
in other cases. (E.g. to execute a back-logged command.) But in such a case
it should not ponder or analyze before processing the command, as these
should only be done while waiting for true input.