Implement periodic updates
The . command now causes printing of a stat01 command, without aborting
the analysis in progress. To achieve this, InputCommand is called during
search when polling reveals there is input. In this case it (currently)
only processes the . command, returning a value that informs the caller
the search should not be aborted. Other commands are backlogged, and
will cause the usual search abort. When InputCommand is then again called
from main it picks up the backlogged command in stead of reading one from
input.