Fix InputCommand patch
There was a done=true in InputCommand that caused only a single command
to be parsed when a command argument was given (and thus also when a
backlogged command was processed). This led to the engine immediately
starting thinking after interruption of ponder, in stead of waiting
for the input move to be processed, with obviously fatal consequences.