Take account of multi-session TC while pondering
We now call RenewTimeControl rather than SetTimeControl to calculate
the new TC parameters when a ponder search produces a move after a
ponder hit, to mimic what would happen if the computer thinks in
the main loop. When accepting the ponder move during search we should
be more careful, because calling the SetTimeControl (indirectly) there
would interfere with the 'time' and 'otim' commands received just
before the move, which already include the time for the next session.
So there we only set the number of moves when a new session starts.
This would not take account of multi-session TCs, but in XBoard mode
there shouldn't be any.