Add time increment after move in XBoard mode
Adding the increment with Fischer clocks was done in SetResponseTime,
so it was included in the time for the current move. The time would
normally already have been added through the 'time' command, however,
so the engine would count itself rich by this, and could be flagged.
In XBoard mode the time increment is now added after the move (just in
case the GUI would not send 'time' commands, which would overrule it
anyway before we have to move next).