Improve calculation of thinking time in sudden-death games
GNU Chess seems to assume a session length of 50 moves in Fischer time
controls, and would be taken by surprise by the fact that it would not
receive new time after that. This is pretty fatal in sudden-death game,
when there is no time increment from which it could do the remaining
moves. So for sudden-death games in XBoard mode it now calculates the
thinking time as if there are at least still 30 moves to go.