WIP Replace seemingly-useless test with an assertion.
On one hand we have macros BlackKing and friends assuming that
PieceList[*][0] are the kings.  OTOH, in a number of places we
check this and do something different if there is something else
in those slots.

Add to this that clang reports potential "use before set" conditions
because of those extra tests.

Pure nonsense. This patch assumes the macro is right, and for
safety's sake adds assertions to detect any discrepancy.

Note to self: check what happens when the board is non-default,
and especially when it is set to an invalid position with no king.
Especially check the tsume mode, since classical tsume to not show
player's king.